+++ src/main/java/roomescape/persistence/dao/JdbcReservationDao.java
@@ -11,7 +11,7 @@ import roomescape.business.domain.PlayTime;
 import roomescape.business.domain.Reservation;
 import roomescape.business.domain.Theme;
 import roomescape.persistence.entity.ReservationEntity;
-import roomescape.presentation.dto.ReservationAvailableTimeResponse;
+import roomescape.presentation.dto.reservation.ReservationAvailableTimeResponse;
 
 @Repository
 public class JdbcReservationDao implements ReservationDao {
