+++ src/main/java/roomescape/presentation/dto/playtime/PlayTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.dto;
+package roomescape.presentation.dto.playtime;
 
 import java.time.LocalTime;
 import roomescape.business.domain.PlayTime;
