+++ src/main/java/roomescape/presentation/dto/reservation/AdminReservationRequest.java
@@ -1,7 +1,6 @@
-package roomescape.presentation.dto.admin;
+package roomescape.presentation.dto.reservation;
 
 import java.time.LocalDate;
-import roomescape.presentation.dto.ReservationRequest;
 
 public record AdminReservationRequest(
         LocalDate date, Long timeId, Long themeId, Long userId
