+++ src/main/java/roomescape/presentation/dto/reservation/ReservationAvailableTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.dto;
+package roomescape.presentation.dto.reservation;
 
 public record ReservationAvailableTimeResponse(String startAt, Long timeId, boolean alreadyBooked) {
 }
