+++ src/test/java/roomescape/fake/FakeReservationDao.java
@@ -11,7 +11,7 @@ import roomescape.persistence.dao.ReservationDao;
 import roomescape.persistence.entity.PlayTimeEntity;
 import roomescape.persistence.entity.ReservationEntity;
 import roomescape.persistence.entity.ThemeEntity;
-import roomescape.presentation.dto.ReservationAvailableTimeResponse;
+import roomescape.presentation.dto.reservation.ReservationAvailableTimeResponse;
 
 public class FakeReservationDao implements ReservationDao {
